+++ sntp/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.5.0] - 2024-06-09
 ### Changed
 - Bumped the minimum `chrono` version from `0.4` to `0.4.32`.
 - Updated `w5500-hl` dependency from `0.11.0` to `0.12.0`.
@@ -44,7 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2022-05-03
 - Initial release
 
-[Unreleased]: https://github.com/newAM/w5500-rs/compare/sntp%2Fv0.4.0...HEAD
+[Unreleased]: https://github.com/newAM/w5500-rs/compare/sntp%2Fv0.5.0...HEAD
+[0.5.0]: https://github.com/newAM/w5500-rs/compare/sntp%2Fv0.4.0...sntp%2Fv0.5.0
 [0.4.0]: https://github.com/newAM/w5500-rs/compare/sntp%2Fv0.3.0...sntp%2Fv0.4.0
 [0.3.0]: https://github.com/newAM/w5500-rs/compare/sntp%2Fv0.2.0...sntp%2Fv0.3.0
 [0.2.0]: https://github.com/newAM/w5500-rs/compare/sntp%2Fv0.1.0...sntp%2Fv0.2.0